backend/requirements.txt
fastapi>=0.110
uvicorn[standard]
numpy
scipy
anthropic>=0.25
pydantic>=2.0
python-dotenv
httpx
